pyproject.toml
[build-system]
requires = [ "poetry-core" ]
build-backend = "poetry.core.masonry.api"

[installer]
no-binary = [ ":none:" ]

[tool.poetry]
name = "awstaga"
version = "0.12.1-pre.0"
description = "Python CLI for tagging AWS resources based on a YAML configuration"
license = "Apache-2.0"
authors = [ "Cliffano Subagio <cliffano@gmail.com>" ]
readme = "README.md"
homepage = "https://github.com/cliffano/awstaga"
repository = "https://github.com/cliffano/awstaga"
documentation = "https://github.com/cliffano/awstaga"
keywords = [ "awstaga", "aws", "tag", "tagging" ]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: Apache Software License",
  "Operating System :: OS Independent"
]
packages = [
    { include = "awstaga" }
]

[tool.poetry.scripts]
awstaga = "awstaga:cli"

[tool.poetry.dependencies]
python = "^3.8"
boto3 = "^1.26.155"
click = "^8.1.3"
conflog = "^1.5.1"
PyYAML = "^6.0.1"
pyyaml-include = "^1.3.1"

[tool.poetry.group.dev.dependencies]
twine = "4.0.2"
wheel = "0.40.0"
yq = "3.2.2"

[tool.poetry.group.test.dependencies]
coverage = "7.2.4"
pylint = "3.0.1"
pytest = "7.3.1"
wily = "1.24.0"

[tool.poetry.group.doc.dependencies]
pylint-report = "2.4.0"
pytest-html = "3.2.0"
Sphinx = "6.2.1"
